Fix weighted post cost build

+++ external/nthu-route/src/router/Congestion.cpp
@@ -74,7 +74,7 @@ double Congestion::get_cost_2d(const Coordinate_2d& c1, const Coordinate_2d& c2,
 
         case MADEOF_COST: {    //Used in part III: Post processing
             if (post_weighted_cost_enabled()) {
-                const int overflow_after_insert = std::max(0, edge.cur_cap + 1 - edge.max_cap);
+                const double overflow_after_insert = std::max(0.0, edge.cur_cap + 1 - edge.max_cap);
                 return overflow_after_insert > 0
                         ? 1.0 + post_weighted_cost_scale() * overflow_after_insert
                         : 0.0;
